AbiTest: Include QDSP6 ELF binaries in validity check
The file system on the device can be used to store executables
and libraries for processors other than the main CPU.  The
architecture of these processors may not be in the list that
AbiTest flags as valid.  Add QDSP6 ELF machine type to the list.

Change-Id: Ica101d0b4ee374e25b63d147bf0501ad751fed76
(cherry picked from commit 65022f93722114bd245b02fe811280a8195415e3)