cts: escaping check for graphics device.
b/12955573

/dev/mm_v3d:
/dev_mm_interlock:

- implements the bcm gpu.  the gpu is used both for composition and for openGL.
openGL acceleration can be initiated from the applications and hence it can
run as part of 'others' group.
- the interlock device is a dummy device that is used for synchronization of any
2 different jobs on the gpu. it does not have access to any hardware registers or
memory.
- the gpu driver uses control list primitives.  those are constructed in userspace
(proprietary library). the list contains information about type of job and
memory addresses where the operations needs to be carried out. it is difficult
(not impossible though) to figure out what is being sent to the hardware.
if a wrong address is programmed, there is a chance that the system will lock up
and hard reset will be required.

/dev/mm_isp:

- this device is used for format conversion via the isp hardware.
- internal format is used for efficiency throughout the system, however if
and when an application needs direct access to the data (eg screen capture)
there is a need to apply a conversion on first access through gralloc.
- since any application can query such data, it can run as part of the
'others' group.

for all of those devices, bcm has reviewed the interfaces in terms of
security and we are confident of the current design.  while it is acknowledged
that there is a slim chance some deny of service situation may arise from
sending invalid information to the hardware.

Change-Id: Ib433430eeaa7cd8f91c0c172da2ff073423d1bd5
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>