De-flake WebView zoom/scale change tests [DO NOT MERGE]
The tests were attempting to zoom the page in immediately after
the loadUrlAndWaitForCompletion method had finished. Because the
PictureListener used in that method is racey the test would
sometimes attempt to zoom the page in too early. Waiting for
canZoomIn to be true fixes the race.

BUG: 15544557
Change-Id: I17ebbecebc1ab35d8afe0c0e6cb72841c59e797b
(cherry picked from commit c91fe5d75d5973c0f5ddfb9cd209eaeb1fc56615)