hardware: consumerir: Increase test pattern length
Increase test pattern length by roughly 10x from ~50 milliseconds to ~500.
Increase test error margin from 10% to 50%.

Devices must be within 50% of the pattern length for transmit, including some
fixed communication latency between the app and hardware.  The point of this
test is to detect devices that are using an incorrect counting scheme (e.g.
periods instead of microseconds) or devices that are wildly off in transmitting
patterns.

This will still reject the cases we want to fail, while allowing for a little
more of a fixed communication latency, for the cost of a slightly longer runtime
for the test.

Bug: 14057146
Change-Id: I4348c1ebb4eeaa5b9ca20f678c0074736b272f63
Signed-off-by: Alex Ray <aray@google.com>