Fix a concurrency bug in OpenSSLHeartbleedTest.
The test was not guaranteed to observe all of the data transmitted by
the client in the scenarios where the client failed. As a result, the
test was occasionally failing (instead of passing) when the client
terminated the TLS/SSL handshake with the fatal alert
unexpected_message.

This CL fixes the test by ensuring that the test observes all of the
data exchanged between the client and the server.

Bug: 13906893

(cherry picked from commit 7e82d7a2011b98f97d5048d2af958a0ec97da74c)

Change-Id: I5091ae7f5e12bdb6fe76a9f0f44618e0e55042c5