Update part of the print CTS tests due to adding print preview.
Now that we have print preview we are asking for the first fifty
pages instead of the first one. Also now the preview is reading
the PDF, so we have to write a blank document instead of doing
nothing. This change fixes the tests in PageRangeAdjustmentTest.

Also this change adds a test for the case when the selected page
is aready written in the batch requested for print preview.

Change-Id: Icfafcb788acef484d9ccb8c05b0c9ed29a8ec335