Fix build breakage due to broken CTS dependencies.
CtsVerifier depended on a static library that was built directly
against the platform, but it depended on the SDK.

This meant that proguard would complain about references in the
static library that were hidden from the SDK (arraycopy(float[]...)
in this case.

This is unsafe in general, but safe in the specific case where
the SDK stubs are generated from the platform that the static library
builds against.

This change compiles CtsVerifier against the platform to fix the build
as a short term workaround.

(cherry picked from commit 8d998397618aed66a3bc448240d2dfc45a7b2ef9)

Change-Id: Ic245e53cb57ca83e0c0053202bb7d12000aed07c