Fix configure check regarding AdaptivePlayback ReconfigDrcTest.
1. Change configureAndStart function type from "void" to "boolean".
When configureAndStart is called, then codec configure and start directly.
But if codec is unsupported(such as not supported high resolution contents),
it generates configure excention. So, it is necessary that check weather
codec is supported or not.

2. Add try-catch in configureAndStart function.
If codec configure failed, then catch this exception. This process can avoid
codec configure exception.

3. Modify configure check condition in ReconfigDrcTest class.
ReconfigDrcTest use 360p and 720p contents. So, if device can't support 720p,
this condition will run next step without exception.

Change-Id: I6b423e74a279d978c43fbebd16f22e67c9cb0f87
Signed-off-by: Joongeun Choi <joongeun.choi@lge.com>