Refactored code to be easier to understand. Added a verifier.
Created packages for the testing infrastructure and the testing
classes. Changed the package name of differencecalculators to
bitmapcomparers, and changed the base class to BitmapComparer.
Added the GoldenImageVerifier to allow for "golden" images
for test case building. For sweeping tests, added logging to
make it easier to see exactly which test failed.

Change-Id: I38d928da5c206e83947917d5949f136b719c8759