Improve CTS generated tests (cts).
Change the way we track the range of allowed values: track
minimum and maximum separately, handle NaN and infinity better,
allow error on input and output.

With these new changes, improve many individual tests.

Instead of setting a global specifying the test expectations,
instantiate an object that contains this info so that parallel
tests won't interfere.

Fix a bug for the rare test that disregards secondary output
when the primary output is NaN.

Change-Id: I6da35c692d6b95a835e2f76f3bb78c7962b09b43