Make LinkifyTest#testAddLinks7 more realistic
One of the tests contained a string of telephone numbers that could
be parsed ambiguously. The parse changed in the update to version 6.0
of libphonenumber, which overall has a higher quality of dealing with
such ambiguous parses (validated against real-world data). This patch
changes the test string to one less likely to be misparsed.

Bug: 17261635
Change-Id: Ib04b979621efc5f1c1476b7765453524de305d9b