Camera2: fix raw test and robustness issues
- testBasicRawCapture is intended to only capture the raw buffer and validate
it, not really intend to validate the metadata, as the test doesn't kick off
3A before capture, the not all 3A related metadata is in the good shape. The
metadata validation is done by other test already.
- Fix the resource release issue in Robustness test. The test relies on GC to
release the native resources associated with the targets. When it finishes one
iteration, and start another iteration, the native resources (mainly gralloc
buffers) are not released quickly, this seems causing the OOM issue randomly
during the test.

Bug: 17508174
Change-Id: Id029ca9342b1802c03e6f28aaf3f4abee0f5d459