Make window introspection tests more robust.
The window introspection CTS tests were hardcoding the assumption
that there is always a nav bar and a status bar which is not always
true. For example, a wearable may not have space for both. This
change makes the tests more rebust by removing tests that rely on
existence of nav and status bars while making other window tests
agnosinc for the existence of these windows.

bug:17499420

Change-Id: Ieabb33ad106f805d49e87169d594f6a0b211864e