MediaCodecTest: don't assume encoder can use input buffers
The test uses createInputSurface() instead of input buffers. According to
the documentation of createInputSurface, it "Requests a Surface to use
as the input to an encoder, _in place of input buffers_." Therefore,
I think it's incorrect to assume that the encoder can hold
num_input_buffers+num_output_buffers. It can only be assumed to hold
num_output_buffers.

This resolves a potential deadlock when calling swapBuffers().

Bug: 17634422
Change-Id: I0fa83c1bf51bcb78c463bdb48b43dec338feb315
Signed-off-by: Matt Gumbel <matthew.k.gumbel@intel.com>