Split SELinux neverallow rules test.
The current neverallow rules test component of the security host tests is
responsible for validating all of the security assertions present as part of
the neverallow rules present, but it offers an all-or-nothing result. Given
the number and scope of those rules, finer granularity is desired.  This CL
dynamically produces a java source file at build-time which has split each
neverallow rule into its own test.

Bug: 18005561

Change-Id: Ib3b454766419532de69f1726bd03215d35d6d495