Avoid calling inflate() multiple times in AnimationDrawableTest
Calling inflate() multiple times has undefined behavior. It has only
worked in this particular test because the specific fields being
tested were consistently overwritten while others were not. It is
extremely unlikely that any developer was relying on this behavior.

This fix is needed to implement theme support in AnimationDrawable
that is consistent with the other drawable classes without breaking
the CTS tests.

Also fixes exception type thrown from ScaleDrawable to be consistent
with other drawable wrapper classes.

BUG: 16045735
Change-Id: Ib36834000e7791585ade0d168bbf011b4d219751