Modified Audio_Output feature usage in CTS tests
Before, many CTS tests were not performed if the audio_output feature
was not enabled. Now, the system will use the default audio stub if
there is no other audio HAL defined, and this stub is mostly
functional, and suitable for CTS tests.

This is needed for audio_output feature support.

bug: 19439530
Change-Id: I8be009969401496edce60ebb021c837da4c370ca