Fix ConcurrentModificationException in sensor related test.
When modify a List when it is in iteration, an exception
"ConcurrentModificationException" will happen. Through method
getCollectedEvents in file TestSensorEventListener.java already
use synchronized and Collections.unmodifiableList, the List
"mCollectedEvents" still can be modified by the object itself.
Need to make a deep copy to the List instead of shollow copy.

Change-Id: Id707a674d910cb1187ae63d48e336b3ec92b02dd
Signed-off-by: Han, He <he.han@intel.com>