Restore CTS_TESTCASES_OUT to before commit a5c060.
Actually in the original change of b46ee5967ae28fdd2279cca2693e304edd7f5c9f
and the cherry-pick change, the values of CTS_TESTCASES_OUT are
different. Not sure why cherry-pick didn't lead to conflict.

It seems there are many more CLs to cherry-pick to use the new value of
CTS_TESTCASES_OUT.
For now just restore the value of CTS_TESTCASES_OUT to fix the build.

Change-Id: I4844668b72839b54c24a9cd07cf42f18f04a3b26