Add tests for C++ vector reflection in RenderScript.
Bug: 22098789

There is a bug in the original implementation of C++ reflection where
vector size is not emitted for any of the types. I have a separate
change to slang that properly allows the vector C++ reflected types to
be emitted.

I also replaced a LOGE with a LOGV, since it is a progress indicator,
and not a failure.

Change-Id: I4dfeee93e7f2547299ae0142f44a39d4efa5c1a4