Provide better diagnostics on CTS failures.
Intializes the arrays to a non-zero value to help diagnose what is causing
the 0 return bug, i.e. is a zero found in b/22556317 because the buffer is
not copied or because it is copied and all null?

Also reports all the failures instead of stopping after the first one.

See b/22556317

Change-Id: I413efc14e67ec548abc2023e8bbb98fd7ccd2eae