CTS: TelephonyProviderTest failing
Bug 21207615

The test doesn't run in the context of being the default SMS app. As
a result, the TelephonyProvider code is never actually executed. The
insert call returns a "fake" result and the test interprets that as
failure. The test is commented out for now, until we figure out a way
to run the test as the default SMS app.

Change-Id: I50e2b006f1ff8950247b7fc3cf41d6356414e21b