CTS test for libmedia OOB write anywhere
Set up a DRM session, spoof the non-secure device as
secure and pass an invalid pointer as the decrypt
destination buffer address which causes a segfault in
mediaserver on devices without the patch.

This test only runs on non-secure (L3) devices, it is
skipped on secure devices since the vulnerability doesn't
exist there.

b/23223325

Change-Id: Ie88a09ffa9ff722d2fb4642c6655c95cc2478eed