Fix RVCVXCheck OpenCV load error and other improvements
Latest release of OpenCV for Android only has v3 library and only
support the corresponding v3.0.0 API for arm64 devices. The v3.0.0 API is no
longer compatible with previous v2.4.9 API. Changes in the verifier code and the
stub library are made for the verifier to resume proper operation.

Other improvements:
  * Added OpenCV Library License display to conform to the BSD license.
  * Added support for the offical report log.
  * Added link to more detailed instruction.

Re-enabled this test case in CTS verifier.

Related bug:
  * b/24193070
  * b/22885758

Change-Id: I36f4f43b259c44e1e8e5f3aea89d29736ea6d4c5