Fix testActionNextAndPreviousAtGranularityLineOverTextExtend FAILS
Target CTS test case :
 android.accessibilityservice.cts.AccessibilityTextTraversalTest
   -- testActionNextAndPreviousAtGranularityLineOverTextExtend

1. Modify layout_width, layout_height, textSize in resource file.
In the case of LDPI devices, dip value will be calculated by *0.75px. (1px = 0.75dpi)
For example, 10dip will be 7.5px and ".5" is not accurate value to display on device.
So, I modify the textSize to 12dip(9px) and in similar propotion, layout_width and layout_height change to 96dip(72px).

2. Change the index in Test Code.
Due to above changes, the other test cases using same resource(R.id.edit) will be fail on the other resolution's devices.
So, I set the index values in following test cases to test properly on the other devices.

 android.accessibilityservice.cts.AccessibilityTextTraversalTest
   -- testActionNextAndPreviousAtGranularityPageOverText
   -- testActionNextAndPreviousAtGranularityPageOverTextExtend

Change-Id: I8da2a42d868a5f4f098ecec84c3f4fc00917d771
Signed-off-by: Seunghyun Kim <mcseunghyun.kim@lge.com>