Fix SeedTest for landscape
By commit ce9362e77a7e71810c0cb8abdef216672d53688d, touch is disabled and
the orientation is locked to portrait during runs of the monkey activity.

After that, by commit 76c263a25f8b9885604f8d73b2e8c0841daece49, touch is
enabled, but the orientation remains locked to portrait during runs of that.

This causes testSeed to be failed when the orientation is landscape.

Change-Id: I8421ddc4f2decb0a80c2533d76bf44495f57f90d
Signed-off-by: Ryoma Sasaki <ryoma.sasaki@brilliantservice.co.jp>