Fix bugs regarding TextView layout size with large font scale in ScrollingText
If a device has a large font scale, testHorizontalMovement(CTS) can fail.
The width of TextView is set to 20dp, but it is not enough for devices which have large font scale.
The width should be increased according to font scale.
So, I added an additional syntax ((int) (LITTLE_SPACE * scaledDensity)) to width of TextView.

Change-Id: I177ef982c61dae0374c598000873b071b1f5a6ee
Signed-off-by: Kang Lee <kang.lee@lge.com>