CTS-5.1_r3 : Fix testBadSurfaceDimensions
Target CTS test case:
android.hardware
android.hardware.camera2.cts.RobustnessTest
-- testBadSurfaceDimensions

Processing the preview size that is set in this test is limited by the size of the display device has been added from CTS-5.1_r3.
As a result of this change, the device equipped with a display device of QVGA or HVGA size has come to Fail this test.
This patch modify weirdSize so that the device equipped with a display device of QVGA size is Pass this test.