Adjust min key size needed for RSASSA-PSS.
This adjusts the Android Keystore CTS test logic which determines
whether RSASSA-PSS can be attempted for a given key size. Previously,
the logic incorrectly assumed that salt size is fixed at 20 bytes
whereas in reality it isn't. The logic also contained an off by one
error: it estimated
  min size = modulus size in bytes - 23 - digest output size in bytes
whereas 22 should've been used instead of 23.

This change fixes the issues by switching to assuming that salt size
will be the same as digest output size, which is the behavior we'd
like to enforce going forward.

Bug: 25770609
Change-Id: Iaa9807841c5780b00d497a7e76eb74469eaf3b3a