Fix testGenerateLayoutParams test
The old test was testing that the result of generateLayoutParams()
was the same as the object passed in. This happened to be the previous
behavior, but was not guaranteed by the docs. A recent change to
the platform creates a new instance based on the old one, which breaks
this test assumption.

The fix is to test that the core width/height values are the same between
the two objects, which is all we should count on.

Issue #25930226 CTs: ViewGroupTest#testGenerateLayoutParams failing

Change-Id: Ia561ad25c2b9fab630c3b81151a801809950aabe