Verify ownership change after package install.
When performing "gift" tests, shift the writing to occur before the
target packages are installed.  This follows the typically pattern
of how OBB files are often written before the actual app is installed.

Then continue with the rest of the tests to verify that gift files
can be read by the newly installed apps.  This verifies that the
FUSE daemon correctly reevaluates the directory and file ownership
after an install occurs.

Also extend gift testing to all package-specific external storage
directories.

Bug: 25399427
Change-Id: Idced83c4b52a89d1cf0d5182e43ef2828aa28f26