Increase delta tolerance in TypefaceTest#testInvalidCmapFont
The TypefaceTest#testInvalidCmapFont test is primarily designed to
crash when the corresponding overflow check is not properly handled.
It also checks the returned metrics, to verify the correct behavior
that the glyphs from an invalid custom font are rejected. The latter
is not a security issue, and some variations can be tolerated. This
patch simply increases the tolerance.

Bug: 26191500
Change-Id: I3e7633caacfe9ed1898778dde013e66436fb1b24