CameraITS: add aspect ratio detection and crop test
Test aspect ratio and check if images are cropped correctly under
each output size.

The test image is a black circle inside a black square. When raw
capture is available, set the height vs. width ratio of the circle
in the full-frame raw as ground truth. Then compare with images of
request combinations of different formats ("jpeg" and "yuv") and
sizes. If raw capture is unavailable, take a picture of the test
image right in front to eliminate shooting angle effect.
The height vs. width ratio for the circle should be close to 1.
Considering shooting position error, aspect ratio greater than 1.05
or smaller than 0.95 will fail the test.

Change-Id: I051890e5aa43e2ffa66e563b989223b03de1571f