Fix NotificationManagerTest.checkNotificationExistence
Symptom:
fail NotificationManagerTest.testCancel().

Root cause:
Initialization of Notification display presence flag is not enough.
As a result, the true is set in the flag at least once during the loop,
then Notification is it will remain true even disappear.

Solution:
The first flag of the loop processing set to false.

Change-Id: If39f9d4199089788036fa32ec59ca29e44114fa7