Fix InputMethodManagerTest#testInputMethodManager.
With the previous commit [1], calling
BaseInputConnection#reportFullscreenMode() from the application side
should have no effect.

This is indeed intentional.  IMM#mFullscreenMode is an implementation
details and so far we do not want applications to be able to send a
fake message to update it.

See also a recent fix in BaseInputConnection#reportFullscreenMode() [2].

  [1]: I571d6cc9c6e461d8994aa7496e7e18be13766411
       159dd47db34151314e2449347251ad32b30e5ea1

  [2]: I43b691ffcf792957a691623dfadf3156c82d8a8b
       79b919c98be0154ae50dac814204300c41fac9b2

Bug: 26910926
Change-Id: Ic402cd2441d9ee752d6a26ea970beea94098611f