Make accessibility cts tests more robust.
AccessibilityManagerTest tests were failing occasionally because
they didn't allow enough time for the services to start. Later
tests also depended on the first one to ensure that services had
started, so when the first test failed, others did as well.

Changing the tests to start the tests inside the tests themselves,
so everything is in one package and we can monitor the services
more closely as they start.

There were also two AndroidTest.xml. It appears that only
OldAndroidTest.xml was used. This change allows a more standard
structure.

Bug: 25850147
Change-Id: I090a7b97f61a1fadc91b4c4e22e478c412554862