Verify accessory mode availability
This change adds a two new checks to USB Accessory Test. First, it
requests that the user puts the phone into charging mode before
begining the test, to make sure that the computer can switch the device
into accessory mode when it is set to charging only. Second, it requests
that the user disconnect and then reconnect the device, to make sure
that accessory mode availability isn't affected by usb resets.

Bug: 26255615
Change-Id: I1b00cd0df48de97310da2e9de50267421a8422da