Run text tests with screenOrientation:nosensor
Make sure the text tests are running on portrait mode. When the system
thinks that the orientation is landscape, since  the activities start in
portrait, they immediately get a configuration change event at the test
start, and another one at the test end. This causes different parts of
the system misbehave since tests are executed pretty fast, around every
~300 milliseconds.

Bug: 27589658
Change-Id: I882c067c1f55c82b8aae13d4d0f50ff1b3b1eb3d