Switch to use the IcuTestRunnerBuilder from external/icu
The IcuTestRunnerBuilder from here was copied into external/icu
and refactored to make it independent of this CTS runner so that
it could be used to run the tests using JUnit on the host. This
switches the CTS IcuTestRunner to use the classes from
external/icu and removes the ones that are no longer needed
here.

Also, adds the root of the CTS coverage tests into the list of
roots to be run.

Bug: 22023363
Change-Id: Ibfe0848fbf215e886615afbbe058b8dd0185a4da