fix testActionNextAndPreviousAtGranularityLineOverTextExtend in ldpi
Modify EditText layout to have same alignment than TextView

In TextExtend test, the EditText underline that shows the editable
property to the user shortens the width of the text content in the widget.
In ldpi, it's not negligible and it fits only 11 characters.

A simple fix, that is implemented in this patch, is to null
the background asset to keep same widget size.

Change-Id: Iedacba8fbcc6627f0b67c68bbb6706ca6cf1209c
Signed-off-by: Nicolas Noury <noury@archos.com>