Run test on UI thread to prevent layout in the middle.
Bug 27747569
The assist structure is being captured and then compared
to the View layout structure. Because the test was being
run on an off-UI thread, it was possible for UI layout to
happen in the middle of the test, causing it to fail.

This CL forces the test to happen on the UI thread to
prevent a layout in the middle of the comparison.

Change-Id: Ibe8f2617d0060ca4c0eb8e8363dcce3b16ceafb8