Shrank refocus test and reference image size
Bug: 27747093

Replaced the test image with a smaller one, so that it can be loaded
into memory on an X86 emulator, or a device with small RAM.

Separated the depthmap from the test image itself, to make it easy to
resize both the image and the depthmap in sync. Adjusted DepthImage
and RGBZ classes to handle a separate image and depthmap.

Adjusted the DepthOfFieldOptions class to make the output image
look reasonable. The blur-at-infinity parameter was improperly set
previously.

Use PSNR for image comparison. Updated the ImageCompare class to do
in-place comparision, instead of loading redundant copies of the images
into memory.

Adjusted MediaStoreSaver class to use a different path to write the
output image, in the case where Pictures directory does not exist,
e.g., on an X86 emulator.

Use PNG instaed of JPG test images, to avoid noises from decompression
on different target platforms.

Change-Id: I281c738780d98bb70404d05b59a430735c546715
(cherry picked from commit b081c55822259a82350a5f813c47bef29296a5ed)