Reducing flakiness from Automater test
The test seemed to be failing because the screen was off or locked.
Added a launchpad activity which will dismiss keyguard and power on the
screen before starting the app_details activity. Also added
remove-active-admin as a tear down command and improved logging
to better troubleshoot failures in the future.

Bug: b/28067602
Change-Id: I348ca8bd4105afa1b37a71959fdadb56b36b76bb