Increase SoundPool loading timeout
Due to separation of the media codec framework code between
processes, loading and decoding of samples started taking more
time, causing the test to timeout.

A quick fix for this is to just increase the timeout.
A long term fix would be to make SoundPool more efficient --
currently it loads all the samples one by one, whereas
the codec framework actually creates a thread for each
sample, so some of them can be actually processed in parallel.

Bug: 28220294
Change-Id: Ia86626f97f019399452ab16572d9db69cd630ab3