Add assertion messages to testReceiveTextMessage.
A great number of SIM cards will not have the phone number available (only the IMSI).
This test relies on the SIM card number to send message to itself.

Added fail with a message to use a suitable SIM card, with available number (which may not be feasible) and rerun the test.

If case the number is available, it is may fail due to no signal.

Added messages to check signal and rerun the test.

Bug: 28311416
Change-Id: I488b9ef1f4306a12ee7712e63bb8e5a6a64d5eaf