Trust CA cert intrinsically in KeyChainTest
CA certs are no longer installed if they are requested at the same
time as a PKCS12 file.

Root cert installation is covered by:
 - CAInstallNotificationVerifierActivity

Root cert usage is covered by:
 - CaCertManagementTest

So it is safe to remove the assumption that the CA is installed, we
don't lose any coverage this way.

Fix: 28423627
Change-Id: I0f5f32a310738ef4e14199b94a19bdc6ed1db9b9