Check bluetooth support when testing "Disable NFC beam"
When testing "Disable NFC beam" a file url to an image it setup to be
transferred via Android Beam. Since a file is to be sent Android
will hand-over the data transfer to bluetooth, but for devices without
bluetooth the transfer will fail.

This patch adds a check so that a large file transfer is performed only
if the device supports bluetooth, if not, a simple text message will be
transferred via Android Beam instead.

Change-Id: Ia867bef1ba24d00e94adb189ef66a5699e1db553