Return all errors, not only the first one.
Refactor this to make it clearer what's going on and to make it easier
to return all of the errors.

Modified the way the mandatory library list is examined to be explicit
about what is being checked.

Change kSystemLibraries to a vector since it's only used to iterate
over.

Bug: 29512261
(cherry picked from commit ce16f1df43c5cce596614df105f6a509a5bdd470)

Change-Id: Icd1b3297c9c77d2a498b9fdb84e49ea9c547d81c